--- a/perf/perftool.py
+++ b/perf/perftool.py
@@ -72,9 +72,8 @@ class Perftool(Test):
         count = 0
         # Built in perf test
         for string in process.run("perf test", ignore_status=True).stderr.splitlines():
-            if 'FAILED' in str(string.splitlines()):
-                self.log.info("Test case failed is %s"
-                              % str(string.splitlines()))
+            if 'FAILED' in string:
+                self.log.info("Test case failed is %s" % string)
 
         # perf testsuite
         for line in build.run_make(self.sourcedir, extra_args='check',